requirements.txt
appdirs==1.4.3
beautifulsoup4==4.7.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
future==0.17.1
gmusicapi==11.1.1
gpsoauth==0.4.1
httplib2==0.12.0
idna==2.8
inflect==2.1.0
irc==17.0
jaraco.classes==2.0
jaraco.collections==2.0
jaraco.functools==2.0
jaraco.itertools==4.0.0
jaraco.logging==2.0
jaraco.stream==2.0
jaraco.text==2.0
lxml==4.2.5
MechanicalSoup==0.11.0
mock==2.0.0
more-itertools==5.0.0
mutagen==1.42.0
oauth2client==4.1.3
pbr==5.1.1
proboscis==1.2.6.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycryptodomex==3.7.2
python-dateutil==2.7.5
python-vlc==3.0.4106
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.6.1
tempora==1.14
urllib3==1.24.1
validictory==1.1.2
